% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAUC.R
\encoding{utf8}
\name{AUC_multiclass}
\alias{AUC_multiclass}
\title{AUC multiclass}
\usage{
AUC_multiclass(obs, pred)
}
\arguments{
\item{obs}{a data.frame with class obervations with n columns and m rows; each row represents a sample, the columns represent the classes of outcomes.}

\item{pred}{a data.frame with class predictions n columns and m rows; each rows represents a sample, the columns represent the classes of outcomes.}
}
\value{
returns a list with the following elements: 
 \itemize{
\item{AUC for n classes}
\item{mAUC}
}
}
\description{
Multinomial models are difficult to assess in across--model way.
}
\details{
method taken from Provost and Domingos 2001, Machine learning
}
\examples{

# generate multinomial probabilties with four classes
pred = gtools::rdirichlet(100, c(0.1,0.1,0.5,0.5)) 

# generate multinomial observations with four classes
obs = t(apply(pred,1,rmultinom,size=1,n=1)) 

AUC_multiclass(obs=obs,pred=pred)

}
\author{
Bob Douma
}

