% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cohen_kappa.R
\encoding{utf8}
\name{cohen_kappa}
\alias{cohen_kappa}
\title{Cohen's Kappa}
\usage{
cohen_kappa(obs, pred)
}
\arguments{
\item{obs}{a data.frame with class obervations with n columns and m rows; each row represents a sample, the columns represent the classes of outcomes.}

\item{pred}{a data.frame with class predictions n columns and m rows; each rows represents a sample, the columns represent the classes of outcomes.}
}
\value{
returns a list with the following elements:
 \itemize{
\item{Kappa}
}
}
\description{
Multinomial models are difficult to assess in across--model way.
}
\details{
Method following Cohen (1960)
}
\examples{
library(gtools)
class = c(1,0,0,0)
obs = t(replicate(100,sample(class)))
pred = t(replicate(100,sample(class)))
cohen_kappa(obs=obs,pred=pred) # calculate kappa
}
\author{
Bob Douma
}

