% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa.multinomial.stats.R
\encoding{utf8}
\name{kappa.multinomial.stats}
\alias{kappa.multinomial.stats}
\title{Kappa multinomial statistics}
\usage{
\method{kappa}{multinomial.stats}(obs, pred, ...)
}
\arguments{
\item{obs}{a dataframe}

\item{pred}{a dataframe}

\item{...}{other arguments to be passed to the underlying functions}
}
\value{
returns a list with the following elements:
 \itemize{
\item{observed} a data.frame with the sum of the class occurences
\item{predicted} a data.frame with the sum of the class probabilities
\item{realized} a data.frame with the sum of the observed class probabilities
\item{pmax} a number with the average probability of the most probable class
}
}
\description{
This function is typically called by kappa_multinomial and gathers the necessary statistics for producing kappa_multinomial
}
\details{


This is details
}
\examples{
NULL
}
\author{
Bob Douma
}

