% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_multinomial.R
\name{kappa_multinomial}
\alias{kappa_multinomial}
\title{kappa_multinomial}
\usage{
kappa_multinomial(obs, pred, ...)
}
\arguments{
\item{obs}{a data.frame with class obervations with n columns and m rows; each row represents a sample, the columns represent the classes of outcomes.}

\item{pred}{a data.frame with class predictions n columns and m rows; each rows represents a sample, the columns represent the classes of outcomes.}
}
\value{
returns a list with the following elements: 
 \itemize{
\item{Kappa_prob}
\item{Kappa_loc}
\item{Kappa_multinomial}
\item{p0}
\item{pe}
\item{pmax}
}
}
\description{
A kappa-like measure for multinomial models
}
\details{
Multinomial models are difficult to assess in across--model way.
}
\examples{
library(gtools)
pred = rdirichlet(100, c(0.1,0.1,0.5,0.5)) # generate multinomial probabilties with four classes
obs = t(apply(pred,1,rmultinom,size=1,n=1)) # generate multinomial observations with four classes
kappa_multinomial(obs=obs,pred=pred) # calculate kappa multinomial




}
\author{
Bob Douma
}

