% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa_multinomial_stats.R
\encoding{utf8}
\name{kappa_multinomial_stats}
\alias{kappa_multinomial_stats}
\title{Kappa multinomial statistics}
\usage{
kappa_multinomial_stats(obs, pred, ...)
}
\arguments{
\item{obs}{a dataframe}

\item{pred}{a dataframe}

\item{...}{other arguments to be passed to the underlying functions}
}
\value{
returns a list with the following elements:
 \itemize{
\item{observed} a data.frame with the sum of the class occurences
\item{predicted} a data.frame with the sum of the class probabilities
\item{realized} a data.frame with the sum of the observed class probabilities
\item{pmax} a number with the average probability of the most probable class
}
}
\description{
This function is typically called by kappa_multinomial and gathers the necessary statistics for producing kappa_multinomial
}
\details{


This is details
}
\examples{
library(gtools)
pred = rdirichlet(100, c(0.1,0.1,0.5,0.5))
obs = t(apply(pred,1,rmultinom,size=1,n=1))
kappa_multinomial_stats(obs=obs,pred=pred)
}
\author{
Bob Douma
}

