% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_binary.R
\encoding{utf8}
\name{prob_to_binary}
\alias{prob_to_binary}
\title{prob_to_binary}
\usage{
prob_to_binary(pred)
}
\arguments{
\item{pred}{a data.frame with class predictions n columns and m rows; each rows represents a sample, the columns represent the classes of outcomes.}
}
\value{
returns a list with the following elements: 

\itemize{
\item{pred}
}
}
\description{
This function transform probabilties to presence/absence, by assuming that the class with highest sample probability if present and the others classes are absent
}
\details{
helper function for the tutorial
}
\examples{

\dontrun{
# generate multinomial probabilties with four classes
pred <- gtools::rdirichlet(100, c(0.1,0.1,0.5,0.5)) 
pred_discrete <- multinomialperform:::prob_to_binary(pred)
obs <- t(apply(pred,1,rmultinom,size=1,n=1)) # generate multinomial observations with four classes
kappa_multinomial(obs=obs,pred=pred_discrete) # calculate kappa multinomial
}

}
\author{
Bob Douma
}

