% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpmultinom.R
\encoding{utf8}
\name{rpmultinom}
\alias{rpmultinom}
\title{Multinomial distribution}
\usage{
rpmultinom(x, skewness, double = FALSE)
}
\arguments{
\item{x}{the number of classes}

\item{skewness}{number to determine the differences between class probabilities, 1 < skewness < 100. The larger skewness the larger the difference between class probabilities}

\item{double}{logical; if TRUE, the difference between he most probable and second most probable class is very small}
}
\value{
returns a list with the following elements:
 \itemize{
\item{rpmultinom} generates random probabilities for x classes
}
}
\description{
This function generates a multinomial probilities for x classes
}
\details{


This is details
}
\examples{
rpmultinom(10,skewness=1,double=FALSE)
}
\author{
Bob Douma
}
\seealso{
package DirichletReg and gtools
}

