% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_max.R
\name{select_max}
\alias{select_max}
\title{select maximum}
\usage{
select_max(x)
}
\arguments{
\item{x}{a vector with sample predictions: the elements represent the classes of outcomes.}
}
\value{
returns a vector with the following elements: 
\itemize{
\item{y}
}
}
\description{
This function select the class with highest probability in a sample. If multiple classes have the same heighest probability then one is randomly chosen
}
\details{
helper function for the tutorial
}
\examples{
library(gtools)
pred = rdirichlet(100, c(0.1,0.1,0.5,0.5)) # generate multinomial probabilties with four classes
pred.bin = multinomialperform:::prob_to_binary(pred)



}
\author{
Bob Douma
}

