% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa_multinomial.R
\encoding{utf8}
\name{kappa_multinomial}
\alias{kappa_multinomial}
\title{A kappa-like measure for multinomial models}
\usage{
kappa_multinomial(x)
}
\arguments{
\item{x}{a list containing data summaries to be used in calculating kappa multinomial. These are usually the result of a call to kappa.multinomial.stats, but can be generated in another fasion.}
}
\value{
returns a list with the following elements:
 \itemize{
\item{Kappa_prob}
\item{Kappa_loc}
\item{Kappa_multinomial}
\item{p0}
\item{pe}
\item{pmax}
}
}
\description{
Multinomial models are difficult to assess in across--model way.
}
\details{
This is details
}
\examples{
NULL
}
\author{
Bob Douma
}

