% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa_multinomial.R
\encoding{utf8}
\name{kappa_multinomial}
\alias{kappa_multinomial}
\title{Performance statistics for multinomial models}
\usage{
kappa_multinomial(i, input, ...)
}
\arguments{
\item{i}{input, probably should generalize this}

\item{input}{figure out logical structure for this}

\item{...}{Additional arguments passed to ?}
}
\value{
returns a list with the following elements:
 \itemize{
\item{Kappa}
\item{some other kappa}
}
}
\description{
Multinomial models are difficult to assess in across--model way.
}
\details{
This is details
}
\examples{
NULL
}
\author{
Bob Douma
}

