% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_multinomial.R
\name{kappa_multinomial}
\alias{kappa_multinomial}
\title{kappa_multinomial}
\usage{
kappa_multinomial(obs, pred, nsim)
}
\arguments{
\item{obs}{a data.frame with class obervations with n columns and m rows; each row represents a sample, the columns represent the classes of outcomes.}

\item{pred}{a data.frame with class predictions n columns and m rows; each rows represents a sample, the columns represent the classes of outcomes.}
\item{nsim}{a number representing the number of randomizations when pe cannot be computed analytically}
}
\value{
returns a list with the following elements: 
 \itemize{
\item{Kappa_prob}
\item{Kappa_loc}
\item{Kappa_multinomial}
\item{p0}
\item{pe}
\item{pmax}
}
}
\description{
A kappa-like measure for multinomial models
}
\details{
Multinomial models are difficult to assess in across--model way.
}
\examples{
# generate multinomial probabilties with four classes
pred = data.frame(gtools::rdirichlet(100, c(0.1,0.1,0.5,0.5)))
# generate multinomial observations with four classes 
obs = data.frame(t(apply(pred,1,rmultinom,size=1,n=1)))
# calculate kappa multinomial 
kappa_multinomial(obs=obs,pred=pred) 

}
\author{
Bob Douma
}

