% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappa_multinomial.R
\encoding{utf8}
\name{kappa_multinomial}
\alias{kappa_multinomial}
\title{A kappa-like measure for multinomial models}
\usage{
kappa_multinomial(x)
}
\arguments{
\item{x}{a list containing data summaries to be used in calculating kappa multinomial. These are usually the result of a call to kappa_multinomial_stats, but can be generated in another fasion.}
}
\value{
returns a list with the following elements:
 \itemize{
\item{Kappa_prob}
\item{Kappa_loc}
\item{Kappa_multinomial}
\item{p0}
\item{pe}
\item{pmax}
}
}
\description{
Multinomial models are difficult to assess in across--model way.
}
\details{
This is details
}
\examples{
library(gtools)
pred = rdirichlet(100, c(0.1,0.1,0.5,0.5))
obs = t(apply(pred,1,rmultinom,size=1,n=1))
k_stats = kappa_multinomial_stats(obs=obs,pred=pred)
kappa_multinomial(k_stats)
}
\author{
Bob Douma
}

